Fixed all Corrock crash when in a modded dimension
Resolves this bug report: https://github.com/minecraftabnormals/The-Endergetic-Expansion/issues/48#issuecomment-626096201

Right now, I wanted to just make the corrock keep their current behavior and only use their overworld form in modded dimensions as default. You can change the getOrDefault to computeIfAbsent and do some sort of check on the dimension to see if it is "end" like or "nether" like to determine variants.

However, I think the better method would be to check the biome that corrock is in for the biome type of NETHER or END for the variants. This can be done by changing isInProperDimension to take the blockpos too and then do CONVERSION.getOrDefault(world.getBiome(blockpos).getCategory(), <default overworld block here>) and have the conversion basically use Biome.Category.THEEND and Biome.Category.NETHER as the key instead.

How you want to proceed is entirely up to you but for right now, this PR just deals with the crash with other mods. Good luck!